chore: move contraint.ts to constant folder
